requirements.txt
wheel
dnspython
pytest
mkdocs
